Add a filter for the WC Shortcode wrapper.
It would be helpful to have some control over the `WC_Shortcode` wrapper. Presently, if we wish to change the wrapping HTML, it involves either placing such changes into the post content (bad!) or creating custom shortcodes with their own wrappers and invoking the WC shortcodes within. The latter is not ideal because it breaks the WC conditionals which detect specific shortcodes in the content.

I added a simple filter for the (presently hard-coded) `$wrapper` variable. I considered adding the `$function` as an additional parameter but don't feel so great about passing a `callable` to a filter. I welcome suggestions on any additional params to pass that can help the user of the filter determine the shortcode being called.